feat: move event/tag/sports methods to GammaClient
Move gamma-related methods from ClobClient:
- get_events, get_event_by_slug, get_event_by_id
- get_tags, get_sports
- parse_gamma_list helper

Refs: #1
